refactor: extract drag drop hooks
